requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
deepl==1.22.0
distro==1.9.0
Django==5.2.3
feedparser==6.0.11
frozenlist==1.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
intel-cmplr-lib-ur==2025.1.1
intel-openmp==2025.1.1
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.22
langchain-teddynote==0.3.45
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
lz4==4.4.4
marshmallow==3.26.1
mkl==2025.1.0
mkl-service==2.4.2
mkl_fft==1.3.14
mkl_random==1.2.10
mmh3==4.1.0
multidict==6.4.4
mypy_extensions==1.1.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==0.28.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pillow==11.2.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.10.0
propcache==0.3.2
protobuf==5.29.5
protoc-gen-openapiv2==0.0.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==78.1.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tavily-python==0.7.5
tbb==2022.1.0
tcmlib==1.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
umf==0.10.0
urllib3==2.4.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
